'Refactor file structure'
Added for faster hot reloading during development.
