+++ client/src/index.js
@@ -1,17 +1,6 @@
 import React from 'react';
-import { Router, Route, IndexRoute } from 'react-router';
 import { render } from 'react-dom';
-
-// Redux
-import { Provider } from 'react-redux';
-import store, { history } from './redux/store';
-
-// Components
-import App from './redux/App';
-import Home from './components/Home';
-import PriceAlert from './components/PriceAlert';
-import Unsubscribe from './components/Unsubscribe';
-import NotFound from './components/NotFound';
+import GamePriceTracker from './GamePriceTracker';
 
 // Service worker disabled for development...
 // import registerServiceWorker from './registerServiceWorker';
@@ -20,25 +9,21 @@ import 'bootstrap/dist/css/bootstrap.css';
 import './index.css';
 
 
-function GamePriceTracker(props) {
-    return (
-        <Provider store={store}>
-            <Router history={history}>
-                <Route path='/' component={App}>
-                    <IndexRoute component={Home} />
-                    <Route path='/games/*' component={PriceAlert} />
-                    <Route path='/manage/*' component={Unsubscribe} />
-                    <Route path='*' component={NotFound} />
-                </Route>
-            </Router>
-        </Provider>
-    );
-}
-
+const rootEl = document.getElementById('root');
 
 render(
     <GamePriceTracker />,
-    document.getElementById('root')
+    rootEl
 );
+
+if (module.hot) {
+    module.hot.accept('./GamePriceTracker', () => {
+        const NextApp = require('./GamePriceTracker').default
+        render(
+            <NextApp />,
+            rootEl
+        );
+    });
+}
 
 // registerServiceWorker();
